Crate problem127

Problem 127 (abc-hits)

Solution detail

Firstly, we'll be computing a lot of radicals. It will save us a lot of time to have a precomputed list of them ready to hand - we only need them for inputs up to the limit of the problem. This list can be computed using a modified Sieve of Eratosthenes.

It will also be convenient to have a list of inputs sorted by radical - the reason will become clear in a moment.

For a given abc-hit, it is true that all of a, b and c are pairwise coprime, so that rad(abc) is equal to rad(a) * rad(b) * rad(c) which saves us some computation. It is also true that rad(a) = rad(abc) / rad(bc) < c / rad(bc) < b / rad(b), where the final inequality is true because c < 2b and rad(c) >= 2.

This leads to the following algorithm. Iterate over all possible b. For each one, iterate over every a with gcd(a, b) = 1 and rad(a) < b / rad(b), and check if we have an abc-hit. The iteration over a is why it is useful to have values sorted by their radical.

Functions

abc_hits
Determine all abc-hits with c less than the given limit.

all_radicals
Compute a list of radicals of all numbers less than the given limit.

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of c in al abc-hits with c less than the limit.